+++ Ametrin/src/main/java/com/ametrinstudios/ametrin/data/DataProviderExtensions.java
@@ -2,10 +2,10 @@ package com.ametrinstudios.ametrin.data;
 
 import net.minecraft.core.registries.BuiltInRegistries;
 import net.minecraft.resources.Identifier;
+import net.minecraft.resources.ResourceKey;
 import net.minecraft.world.item.Item;
-import net.minecraft.world.item.Items;
+import net.minecraft.world.level.ItemLike;
 import net.minecraft.world.level.block.Block;
-import net.minecraft.world.level.block.Blocks;
 
 import java.util.HashSet;
 
@@ -52,15 +52,22 @@ public final class DataProviderExtensions {
     }
 
     public static Identifier getItemKey(Item item) {
-        var key = BuiltInRegistries.ITEM.getKey(item);
-        if(key.getPath().equals("air") && key.getNamespace().equals("minecraft") && item != Items.AIR) throw new IllegalArgumentException("item " + item + "not found");
-        return key;
-
+        return getItemResourceKey(item).identifier();
     }
 
     public static Identifier getBlockKey(Block block) {
-        var key = BuiltInRegistries.BLOCK.getKey(block);
-        if(key.getPath().equals("air") && key.getNamespace().equals("minecraft") && block != Blocks.AIR) throw new IllegalArgumentException("block " + block + "not found");
-        return key;
+        return getBlockResouceKey(block).identifier();
+    }
+
+    public static ResourceKey<Item> getItemResourceKey(Item item) {
+        return BuiltInRegistries.ITEM.getResourceKey(item).orElseThrow();
+    }
+
+    public static ResourceKey<Item> getItemResourceKey(ItemLike item) {
+        return getItemResourceKey(item.asItem());
+    }
+
+    public static ResourceKey<Block> getBlockResouceKey(Block block) {
+        return BuiltInRegistries.BLOCK.getResourceKey(block).orElseThrow();
     }
 }