+++ Ametrin/src/main/java/com/ametrinstudios/ametrin/util/mixin/IExtendedBlockModelGenerators.java
@@ -1,0 +1,17 @@
+package com.ametrinstudios.ametrin.util.mixin;
+
+import net.minecraft.client.data.models.BlockModelGenerators;
+import net.minecraft.client.data.models.model.TextureMapping;
+import net.minecraft.world.level.block.Block;
+
+import java.util.function.Function;
+
+public interface IExtendedBlockModelGenerators {
+    default BlockModelGenerators.BlockFamilyProvider familyWithExistingFullBlock(Block fullBlock, TextureMapping mapping) {
+        throw new AssertionError("Implemented via Mixin");
+    }
+
+    default BlockModelGenerators.BlockFamilyProvider familyWithExistingFullBlock(Block fullBlock, Function<Block, TextureMapping> mapping) {
+        throw new AssertionError("Implemented via Mixin");
+    }
+}